+++ Tests/AdvancedMappingCrud.DbContext.Tests/AdvancedMappingCrud.DbContext.Tests.IntegrationTests/Tests/GiftCards/CreateGiftCardTests.cs
@@ -1,0 +1,16 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.AspNetCore.IntegrationTesting.ServiceEndpointTest", Version = "1.0")]
+
+namespace AdvancedMappingCrud.DbContext.Tests.IntegrationTests.Tests.GiftCards
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    [Collection("SharedContainer")]
+    public class CreateGiftCardTests : BaseIntegrationTest
+    {
+        public CreateGiftCardTests(IntegrationTestWebAppFactory factory) : base(factory)
+        {
+        }
+    }
+}