+++ Tests/AdvancedMappingCrud.Repositories.Tests/AdvancedMappingCrud.Repositories.Tests.IntegrationTests/Tests/GiftCards/GetGiftCardByIdTests.cs
@@ -1,0 +1,16 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.AspNetCore.IntegrationTesting.ServiceEndpointTest", Version = "1.0")]
+
+namespace AdvancedMappingCrud.Repositories.Tests.IntegrationTests.Tests.GiftCards
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    [Collection("SharedContainer")]
+    public class GetGiftCardByIdTests : BaseIntegrationTest
+    {
+        public GetGiftCardByIdTests(IntegrationTestWebAppFactory factory) : base(factory)
+        {
+        }
+    }
+}