+++ Modules/Intent.Modules.Application.DomainInteractions/release-notes.md
@@ -1,3 +1,6 @@
+### Version 1.2.7
+- Fixed: When a user-supplied primary key is used, use its name in the `return` statement instead of always assuming it is "id" 
+
 ### Version 1.2.6
 - Improvement: Updated code generation for query handlers when Mapperly is used to avoid premature materializing collections before mapping. 
 - Fixed: When two or more DTOs in the same folder are of the form `<Entity>Dto.cs` and `<Entity><SomeString>Dto.cs` the correct DTO is now injected into the constructor of the corresponding handlers